+++ protocol/csharp/gate.cs
@@ -242,7 +242,7 @@ namespace abelkhan
 
     public class client_call_gate_caller : abelkhan.Icaller {
         public static client_call_gate_rsp_cb rsp_cb_client_call_gate_handle = null;
-        private Int64 uuid_2a41ded1_acf2_3b8c_95bc_f149a01703b2 = (Int64)RandomUUID.random();
+        private Int32 uuid_2a41ded1_acf2_3b8c_95bc_f149a01703b2 = (Int32)RandomUUID.random();
 
         public client_call_gate_caller(abelkhan.Ichannel _ch, abelkhan.modulemng modules) : base("client_call_gate", _ch)
         {
@@ -253,7 +253,7 @@ namespace abelkhan
         }
 
         public client_call_gate_heartbeats_cb heartbeats(){
-            var uuid_a514ca5f_2c67_5668_aac0_354397bdce36 = (UInt64)Interlocked.Increment(ref uuid_2a41ded1_acf2_3b8c_95bc_f149a01703b2);
+            var uuid_a514ca5f_2c67_5668_aac0_354397bdce36 = (UInt32)Interlocked.Increment(ref uuid_2a41ded1_acf2_3b8c_95bc_f149a01703b2);
 
             var _argv_6fbd85be_a054_37ed_b3ea_cced2f90fda4 = new ArrayList();
             _argv_6fbd85be_a054_37ed_b3ea_cced2f90fda4.Add(uuid_a514ca5f_2c67_5668_aac0_354397bdce36);
@@ -268,7 +268,7 @@ namespace abelkhan
         }
 
         public client_call_gate_get_hub_info_cb get_hub_info(string hub_type){
-            var uuid_e9d2753f_7d38_512d_80ff_7aae13508048 = (UInt64)Interlocked.Increment(ref uuid_2a41ded1_acf2_3b8c_95bc_f149a01703b2);
+            var uuid_e9d2753f_7d38_512d_80ff_7aae13508048 = (UInt32)Interlocked.Increment(ref uuid_2a41ded1_acf2_3b8c_95bc_f149a01703b2);
 
             var _argv_64f76bda_d44d_3aed_a6a4_d85fea361e24 = new ArrayList();
             _argv_64f76bda_d44d_3aed_a6a4_d85fea361e24.Add(uuid_e9d2753f_7d38_512d_80ff_7aae13508048);
@@ -494,7 +494,7 @@ namespace abelkhan
 
     public class hub_call_gate_caller : abelkhan.Icaller {
         public static hub_call_gate_rsp_cb rsp_cb_hub_call_gate_handle = null;
-        private Int64 uuid_9796175c_1119_3833_bf31_5ee139b40edc = (Int64)RandomUUID.random();
+        private Int32 uuid_9796175c_1119_3833_bf31_5ee139b40edc = (Int32)RandomUUID.random();
 
         public hub_call_gate_caller(abelkhan.Ichannel _ch, abelkhan.modulemng modules) : base("hub_call_gate", _ch)
         {
@@ -505,7 +505,7 @@ namespace abelkhan
         }
 
         public hub_call_gate_reg_hub_cb reg_hub(string hub_name, string hub_type){
-            var uuid_98c51fef_38ce_530a_b8e9_1adcd50b1106 = (UInt64)Interlocked.Increment(ref uuid_9796175c_1119_3833_bf31_5ee139b40edc);
+            var uuid_98c51fef_38ce_530a_b8e9_1adcd50b1106 = (UInt32)Interlocked.Increment(ref uuid_9796175c_1119_3833_bf31_5ee139b40edc);
 
             var _argv_e096e269_1e08_36d1_9ba4_b7db8c8ff8a7 = new ArrayList();
             _argv_e096e269_1e08_36d1_9ba4_b7db8c8ff8a7.Add(uuid_98c51fef_38ce_530a_b8e9_1adcd50b1106);
@@ -522,7 +522,7 @@ namespace abelkhan
         }
 
         public hub_call_gate_reverse_reg_client_hub_cb reverse_reg_client_hub(string client_uuid){
-            var uuid_5352b179_7aef_5875_a08f_06381972529f = (UInt64)Interlocked.Increment(ref uuid_9796175c_1119_3833_bf31_5ee139b40edc);
+            var uuid_5352b179_7aef_5875_a08f_06381972529f = (UInt32)Interlocked.Increment(ref uuid_9796175c_1119_3833_bf31_5ee139b40edc);
 
             var _argv_03d844bd_f79a_3179_8f8b_9f0ed380f60c = new ArrayList();
             _argv_03d844bd_f79a_3179_8f8b_9f0ed380f60c.Add(uuid_5352b179_7aef_5875_a08f_06381972529f);
